Cart API complete

+++ store/views.py
@@ -1,32 +1,15 @@
+from store.pagination import DefaultPagination
+from django.db.models.aggregates import Count
+from django.shortcuts import get_object_or_404
 from django_filters.rest_framework import DjangoFilterBackend
-from rest_framework.response import Response
-from rest_framework import status
-from rest_framework.viewsets import ModelViewSet, GenericViewSet
-from .serializer import (
-    AddCartItemSerializer,
-    CartItemSerializer,
-    CartSerializer,
-    CollectionSerializer,
-    ProductSerializer,
-    ReviewSerializer,
-    UpdateCartItemSerializer,
-)
-
-from rest_framework.mixins import (
-    CreateModelMixin,
-    RetrieveModelMixin,
-    DestroyModelMixin,
-)
 from rest_framework.filters import SearchFilter, OrderingFilter
-from .serializer import (
-    ProductSerializer,
-    CollectionSerializer,
-    ReviewSerializer,
-    CartSerializer,
-    CartItemSerializer,
-)
-from .models import Product, Collection, Cart, CartItem
+from rest_framework.mixins import CreateModelMixin, DestroyModelMixin, RetrieveModelMixin
+from rest_framework.response import Response
+from rest_framework.viewsets import ModelViewSet, GenericViewSet
+from rest_framework import status
 from .filters import ProductFilter
+from .models import Cart, CartItem, Collection, Product, Review
+from .serializer import AddCartItemSerializer, CartItemSerializer, CartSerializer, CollectionSerializer, ProductSerializer, ReviewSerializer, UpdateCartItemSerializer
 
 
 class ProductViewSet(ModelViewSet):
@@ -34,69 +17,66 @@ class ProductViewSet(ModelViewSet):
     serializer_class = ProductSerializer
     filter_backends = [DjangoFilterBackend, SearchFilter, OrderingFilter]
     filterset_class = ProductFilter
-    search_fields = ["title", "description"]
-    ordering_fields = ["unit_price", "last_update"]
+    pagination_class = DefaultPagination
+    search_fields = ['title', 'description']
+    ordering_fields = ['unit_price', 'last_update']
 
     def get_serializer_context(self):
-        return {"request": self.request}
+        return {'request': self.request}
 
-    def destroy(self, request, pk, *args, **kwargs):
-        try:
-            product = Product.objects.get(pk=pk)
-            product.delete()
-            return Response(
-                "Product deleted successfully", status=status.HTTP_204_NO_CONTENT
-            )
-        except Product.DoesNotExist:
-            return Response("Product not found", status=status.HTTP_404_NOT_FOUND)
+    def delete(self, request, pk):
+        product = get_object_or_404(Product, pk=pk)
+        if product.orderitems.count() > 0:
+            return Response({'error': 'Product cannot be deleted because it is associated with an order item.'}, status=status.HTTP_405_METHOD_NOT_ALLOWED)
+        product.delete()
+        return Response(status=status.HTTP_204_NO_CONTENT)
 
 
 class CollectionViewSet(ModelViewSet):
-    queryset = Collection.objects.all()
+    queryset = Collection.objects.annotate(
+        products_count=Count('products')).all()
     serializer_class = CollectionSerializer
 
-    def get_serializer_context(self):
-        return {"request": self.request}
-
     def delete(self, request, pk):
-        collection = Collection.objects.get(pk=pk)
+        collection = get_object_or_404(Collection, pk=pk)
+        if collection.products.count() > 0:
+            return Response({'error': 'Collection cannot be deleted because it includes one or more products.'}, status=status.HTTP_405_METHOD_NOT_ALLOWED)
         collection.delete()
-        return Response(
-            "Collection deleted successfully", status=status.HTTP_204_NO_CONTENT
-        )
+        return Response(status=status.HTTP_204_NO_CONTENT)
 
 
 class ReviewViewSet(ModelViewSet):
     serializer_class = ReviewSerializer
 
     def get_queryset(self):
-        return Product.objects.get(pk=self.kwargs["product_pk"]).reviews.all()
+        return Review.objects.filter(product_id=self.kwargs['product_pk'])
 
     def get_serializer_context(self):
-        return {"product_id": self.kwargs["product_pk"]}
+        return {'product_id': self.kwargs['product_pk']}
 
 
-class CartViewSet(
-    CreateModelMixin, RetrieveModelMixin, DestroyModelMixin, GenericViewSet
-):
-    queryset = Cart.objects.prefetch_related("items__product").all()
+class CartViewSet(CreateModelMixin,
+                  RetrieveModelMixin,
+                  DestroyModelMixin,
+                  GenericViewSet):
+    queryset = Cart.objects.prefetch_related('items__product').all()
     serializer_class = CartSerializer
 
 
 class CartItemViewSet(ModelViewSet):
-    http_method_names = ["get", "post", "patch", "delete"]
-
+    http_method_names = ['get', 'post', 'patch', 'delete']
+   
     def get_serializer_class(self):
-        if self.request.method == "POST":
+        if self.request.method == 'POST':
             return AddCartItemSerializer
-        elif self.request.method == "PATCH":
+        elif self.request.method == 'PATCH':
             return UpdateCartItemSerializer
         return CartItemSerializer
 
     def get_serializer_context(self):
-        return {"cart_id": self.kwargs["cart_pk"]}
+        return {'cart_id': self.kwargs['cart_pk']}
 
     def get_queryset(self):
-        return CartItem.objects.filter(cart_id=self.kwargs["cart_pk"]).select_related(
-            "product"
-        )
+        return CartItem.objects \
+                .filter(cart_id=self.kwargs['cart_pk']) \
+                .select_related('product')
